Decouple the gbz80 instruction implementation
