+++ cpu.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "core.h"
+
+void gb_run_frame(struct gb_s *gb);
